Logic for Database Add/RM

+++ app/areas/mutations/deleteArea.ts
@@ -3,12 +3,20 @@ import db from "db"
 import { z } from "zod"
 
 const DeleteArea = z.object({
+  badZip: z.number(),
   id: z.number(),
 })
 
-export default resolver.pipe(resolver.zod(DeleteArea), resolver.authorize(), async ({ id }) => {
-  // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-  const area = await db.area.deleteMany({ where: { id } })
+export default resolver.pipe(
+  resolver.zod(DeleteArea),
+  resolver.authorize(),
+  async ({ badZip, id }) => {
+    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
+    const user = await db.user.findFirst({ where: { id: id } })
+    if (!user) return
 
-  return area
-})
+    const area = await db.area.deleteMany({ where: { zipcode: badZip, userId: user.id } })
+
+    return area
+  }
+)
